pyproject.toml
[project]
name = "nhentai2pdf"
version = "0.1.0"
description = "A high-performance, asynchronous Python utility designed to fetch, normalize, and archive galleries into a structured PDF format."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.11.13",
    "beautifulsoup4>=4.13.3",
    "cloudscraper>=1.2.71",
    "pikepdf>=9.5.2",
    "pillow>=11.1.0",
    "tqdm>=4.67.1",
    "img2pdf>=0.6.3",
    "lxml>=6.0.2",
]

[project.scripts]
nh2pdf = "nhentai2pdf:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
